+++ src/workprint/models/__init__.py
@@ -1,5 +1,0 @@
-from .message import NormalizedMessage
-from .observation import Observation
-from .investigation import Investigation
-
-__all__ = ["NormalizedMessage", "Observation", "Investigation"]